#276: vrt: move function traits to utils
